refactor: update build config for @primer/doc-gen to rolldown (#8041)
