+++ packages/doc-gen/package.json
@@ -21,32 +21,23 @@
   ],
   "scripts": {
     "clean": "rimraf dist",
-    "build": "rollup -c",
+    "build": "rolldown -c",
     "type-check": "tsc --noEmit",
-    "watch": "rollup -c -w",
+    "watch": "rolldown -c -w",
     "test": "vitest --run",
     "test:watch": "vitest"
   },
   "dependencies": {
-    "@babel/runtime": "^7.28.6",
     "@types/signale": "^1.4.7",
     "@types/yargs": "^17.0.33",
     "prettier": "^3.4.2",
     "signale": "^1.4.0",
+    "typescript": "^6.0.3",
     "yargs": "^18.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.29.0",
-    "@babel/plugin-transform-runtime": "^7.29.0",
-    "@babel/preset-env": "^7.29.0",
-    "@babel/preset-typescript": "^7.28.5",
-    "@rollup/plugin-babel": "^6.1.0",
-    "@rollup/plugin-commonjs": "^29.0.0",
-    "@rollup/plugin-json": "^6.1.0",
-    "@rollup/plugin-node-resolve": "^16.0.3",
     "rimraf": "^6.0.1",
-    "rollup": "^4.59.0",
-    "rollup-plugin-typescript2": "^0.36.0",
-    "typescript": "^6.0.3"
+    "rolldown": "^1.1.2",
+    "rolldown-plugin-dts": "^0.26.0"
   }
 }
